PresenceLight/requirements.txt
psutil==5.9.5
rumps==0.4.0
flask==2.3.3
pyserial==3.5
requests==2.31.0
pynput==1.7.6